+++ rdagent/scenarios/data_science/proposal/exp_gen/base.py
@@ -3,7 +3,6 @@ from typing import List, Literal
 
 from rdagent.app.data_science.conf import DS_RD_SETTING
 from rdagent.core.evolving_framework import KnowledgeBase
-from rdagent.core.experiment import ASpecificExp, Experiment
 from rdagent.core.proposal import ExperimentFeedback, Hypothesis, Trace
 from rdagent.scenarios.data_science.experiment.experiment import COMPONENT, DSExperiment
 from rdagent.scenarios.data_science.scen import DataScienceScen
@@ -62,20 +61,12 @@ class DSTrace(Trace[DataScienceScen, KnowledgeBase]):
 
         self.knowledge_base = knowledge_base
 
-        self.current_selection: tuple[int, ...] = (-1,)
-
         self.sota_exp_to_submit: DSExperiment | None = None  # grab the global best exp to submit
 
     COMPLETE_ORDER = ("DataLoadSpec", "FeatureEng", "Model", "Ensemble", "Workflow")
 
     def set_sota_exp_to_submit(self, exp: DSExperiment) -> None:
         self.sota_exp_to_submit = exp
-
-    def get_current_selection(self) -> tuple[int, ...]:
-        return self.current_selection
-
-    def set_current_selection(self, selection: tuple[int, ...]) -> None:
-        self.current_selection = selection
 
     @property
     def sub_trace_count(self) -> int:
@@ -149,76 +140,6 @@ class DSTrace(Trace[DataScienceScen, KnowledgeBase]):
 
         else:
             raise ValueError(f"Invalid search type: {search_type}")
-
-    def is_selection_new_tree(self, selection: tuple[int, ...] | None = None) -> bool:
-        """
-        Check if the current trace is a new tree.
-        - selection maybe (-1,) when the dag_parent is empty.
-        """
-        if selection is None:
-            selection = self.get_current_selection()
-
-        if selection == self.NEW_ROOT or len(self.dag_parent) == 0:
-            return True
-
-        return False
-
-    def collect_all_ancestors(
-        self,
-        selection: tuple[int, ...] | None = None,
-    ) -> list[tuple[DSExperiment, ExperimentFeedback]]:
-        """
-        Collect all ancestors of the given selection.
-        The return list follows the order of [root->...->parent->current_node].
-        """
-        if selection is None:
-            selection = self.get_current_selection()
-
-        if self.is_selection_new_tree(selection):
-            return []
-
-        else:
-            all_ancestors = []
-
-            # start from the latest selection
-            current_node_idx = selection[0]
-
-            # add the current node to the list
-            all_ancestors.insert(0, self.hist[current_node_idx])
-
-            parent_idx = self.dag_parent[current_node_idx]
-
-            while len(parent_idx) > 0:
-                all_ancestors.insert(0, self.hist[parent_idx[0]])
-                parent_idx = self.dag_parent[parent_idx[0]]
-
-        return all_ancestors
-
-    def exp2idx(self, exp: Experiment | List[Experiment]) -> int | List[int]:
-        single_input = not isinstance(exp, list)
-        exps = [exp] if single_input else exp
-        idxs = [i for i, (_exp, _) in enumerate(self.hist) if _exp in exps]
-        if single_input:
-            return idxs[0]
-        else:
-            return idxs
-
-    def idx2exp(self, idx: int | List[int]) -> Experiment | List[Experiment]:
-        single_input = not isinstance(idx, list)
-        idxs = [idx] if single_input else idx
-        if single_input:
-            return self.hist[idxs[0]][0]
-        else:
-            return [self.hist[_idx][0] for _idx in idxs]
-
-    def is_parent(self, parent_idx: int, child_idx: int) -> bool:
-        ancestors = self.get_parents(child_idx)
-        return parent_idx in ancestors
-
-    def get_parents(self, child_idx: int) -> List[int]:
-        ancestors = self.collect_all_ancestors((child_idx,))
-        ancestor_exps = [exp for exp, _ in ancestors]
-        return self.exp2idx(ancestor_exps)
 
     def next_incomplete_component(
         self,
